Refactor : Westagram 리팩토링 완료
