+++ src/Pages/JIYUNLEE/Main/Feeds/SecondFeed.js
@@ -1,65 +1,14 @@
 import React, { Component } from "react";
+import ArticleContents from "./ArticleContents";
 import "./Feeds.scss";
 
 class Feed extends Component {
-  constructor() {
-    super();
-    this.state = {
-      commentsList: [
-        { id: 0, text: "댓글 Component를 분리하고 싶었는데..." },
-        { id: 1, text: "오늘까지는 무리 흑흑흑" },
-      ],
-      id: 2,
-      textarea: "",
-    };
-  }
-
-  handleChangeComment = (e) => {
-    this.setState({ textarea: e.target.value });
-  };
-
-  handleCreateComment = () => {
-    const { textarea, commentsList, id } = this.state;
-    this.setState({
-      id: id + 1,
-      commentsList: [...commentsList, { id: id, text: textarea }],
-      textarea: "",
-    });
-  };
-
-  handleEnterSubmit = (e) => {
-    if (e.key === "Enter") {
-      this.handleCreateComment();
-    }
-  };
-
-  handleClickSubmit = () => {
-    this.handleCreateComment();
-  };
-
   handleChangeLikes = (e) => {
     e.target.setAttribute("src", "/Images/JIYUNLEE/redheart.gif");
   };
 
-  handleRemoveComent = (e) => {
-    const { commentsList } = this.state;
-    let deleteTarget = e.target.parentNode.parentNode.id;
-    const result = commentsList.filter((i) => i.id !== Number(deleteTarget));
-    this.setState({
-      commentsList: result,
-    });
-  };
-
   render() {
-    const {
-      textarea,
-      handleChangeComment,
-      handleClickSubmit,
-      handleEnterSubmit,
-      handleRemoveComent,
-      handleChangeLikes,
-    } = this;
-    const isEnabled = this.state.textarea.length > 0;
+    const { handleChangeLikes } = this;
 
     return (
       <article>
@@ -104,84 +53,7 @@ class Feed extends Component {
             src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/bookmark.png"
           ></img>
         </div>
-
-        <div className="articleContent">
-          <div className="likes">
-            <img className="likepic" alt="" src="/images/JIYUNLEE/robert.jpg" />
-            <span className="likeId">
-              <strong>robertdowneyjr</strong>
-            </span>
-            <span>
-              님 <strong>외 3명</strong>이 좋아합니다.
-            </span>
-          </div>
-          <div className="feedContents">
-            <span>
-              <strong>limosum91</strong>
-            </span>
-            <div className="texts">
-              <span className="notOverFlow"></span>
-              <span className="paragraph">
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                Phasellus imperdiet, nulla et dictum interdum, nisi lorem
-                egestas vitae scelerisque enim ligula venenatis dolor. Maecenas
-                nisl est, ultrices nec congue eget, auctor vitae massa. Fusce
-                luctus vestibulum augue ut aliquet. Nunc sagittis dictum nisi,
-                sed ullamcorper ipsum dignissim ac. In at libero sed nunc
-                venenatis imperdiet sed ornare turpis. Donec vitae dui eget
-                tellus gravida venenatis. Integer fringilla congue eros non
-                fermentum. Sed dapibus pulvinar nibh tempor porta.
-              </span>
-              <button className="moreBtn">더 보기</button>
-            </div>
-          </div>
-          <div className="showComments">
-            <ul className="commentsList">
-              {this.state.commentsList.map((text) => {
-                return (
-                  <li id={text.id}>
-                    <div className="newComment">
-                      <strong>limosum91 </strong>
-                      {text.text}
-                    </div>
-                    <div className="newCommentBtn">
-                      <img
-                        type="button"
-                        className="commentDelete"
-                        alt=""
-                        src="Images/JIYUNLEE/xBtn.png"
-                        onClick={handleRemoveComent}
-                      ></img>
-                      <img
-                        className="commentLikes"
-                        alt=""
-                        src="/Images/JIYUNLEE/empty.png"
-                      ></img>
-                    </div>
-                  </li>
-                );
-              })}
-            </ul>
-          </div>
-        </div>
-        <div className="comments">
-          <div className="createComment" autoComplete="off">
-            <textarea
-              className="inputComment"
-              placeholder="댓글 달기..."
-              onKeyUp={handleEnterSubmit}
-              onChange={handleChangeComment}
-              value={this.state.textarea}
-            ></textarea>
-            <button
-              className={isEnabled ? "activateBtn" : "btnNotAvailable"}
-              disabled={!this.state.textarea}
-              onClick={handleClickSubmit}
-            >
-              게시
-            </button>
-          </div>
-        </div>
+        <ArticleContents />
       </article>
     );
   }
